e-track/requirements.txt
requests>=2.0
psycopg2-binary>=2.9
python-dotenv>=0.21
Flask>=2.2
APScheduler>=3.9
